==> dev/00_prep_corpus.ipynb <==
import os
from os import path
import shutil
import pandas as pd
import random
from tqdm.notebook import tqdm

df_corpus = pd.read_csv("/home/crazyjeannot/Documents/doctorat/data/data_chapitres/metadata/corpus.csv")
path_to_txt = "/home/crazyjeannot/Documents/doctorat/corpora/corpus_chapitres_txt/"
path_to_chunks_txt = "/home/crazyjeannot/Documents/doctorat/corpora/corpus_chapitres_chunks_txt/"

df_corpus.head(1)

def select_and_copy_novels(N, path_to_folder, metadata_df):
    # Create a new folder to store selected novels
    new_folder_path = os.path.join(os.getcwd(), '../corpus/random_txt')
    os.makedirs(new_folder_path, exist_ok=True)

    # Filter novels based on subgenre
    filtered_novels = metadata_df[(metadata_df['genre'] != "roman d'aventures") & (metadata_df['genre'] != "roman historique") & (metadata_df['date'] >= 1840) & (metadata_df['date'] <= 1935)]

    # Randomly select N novels
    selected_novels = random.sample(filtered_novels['doc_name'].tolist(), N)

    # Copy selected novels to the new folder
    for novel in selected_novels:
        novel_path = os.path.join(path_to_folder, f"{novel}.txt")
        destination_path = os.path.join(new_folder_path, f"{novel}.txt")
        shutil.copy2(novel_path, destination_path)

    return selected_novels, new_folder_path

selected_novels, new_folder_path = select_and_copy_novels(102, path_to_txt, df_corpus)

def create_metadata_dataframe(path_to_aventures, path_to_random):
    aventures_files = [f.replace('.txt', '') for f in os.listdir(path_to_aventures) if f.endswith('.txt')]
    random_files = [f.replace('.txt', '') for f in os.listdir(path_to_random) if f.endswith('.txt')]

    aventures_df = pd.DataFrame({'doc_name': aventures_files, 'subgenre': 'aventure'})
    random_df = pd.DataFrame({'doc_name': random_files, 'subgenre': 'random'})

    metadata_df = pd.concat([aventures_df, random_df], ignore_index=True)

    # Save the DataFrame to a CSV file
    metadata_df.to_csv('../data/metadata/metadata.csv', index=False)

    return metadata_df

df_meta = create_metadata_dataframe("/home/crazyjeannot/Documents/doctorat/AVENTURES/corpus/aventures_txt/", "/home/crazyjeannot/Documents/doctorat/AVENTURES/corpus/random_txt/") 

# Chunkise Texts

def chunkize_text_files(N, M, path_to_folder, new_dir):
    # Create a new directory if it doesn't exist
    os.makedirs(new_dir, exist_ok=True)

    # Iterate through each text file in the given folder
    for filename in tqdm(os.listdir(path_to_folder), total=len(os.listdir(path_to_folder))):
        if filename.endswith(".txt"):
            file_path = os.path.join(path_to_folder, filename)
            with open(file_path, 'r', encoding='utf-8') as file:
                content = file.read()
                tokens = content.split()

                # Create overlapping chunks
                for i in range(0, len(tokens) - N + 1, M):
                    chunk = tokens[i:i + N]

                    # Save the chunk to a new text file
                    chunk_filename = f"{filename}_chunk_{i // M}.txt"
                    chunk_filepath = os.path.join(new_dir, chunk_filename)
                    with open(chunk_filepath, 'w', encoding='utf-8') as chunk_file:
                        chunk_file.write(' '.join(chunk))
                        

chunkize_text_files(3000, 1000, path_to_txt, path_to_chunks_txt)

# SELECT N RANDOM CHUNKS

path_to_N_random_chunks = "/home/crazyjeannot/Documents/doctorat/AVENTURES/data/10000_random_chapitres_chunks_txt/"

def random_copy_files(source_folder, destination_folder, num_files=10000):
    # Get a list of all files in the source folder
    all_files = os.listdir(source_folder)

    # Ensure that the number of files to copy is not greater than the total number of files in the source folder
    num_files = min(num_files, len(all_files))

    # Randomly select num_files files
    selected_files = random.sample(all_files, num_files)

    # Create the destination folder if it doesn't exist
    if not os.path.exists(destination_folder):
        os.makedirs(destination_folder)

    # Copy selected files to the destination folder
    for file_name in selected_files:
        source_path = os.path.join(source_folder, file_name)
        destination_path = os.path.join(destination_folder, file_name)
        shutil.copy2(source_path, destination_path)


random_copy_files(path_to_chunks_txt, path_to_N_random_chunks, 10000)

## Delete selected Files

df = pd.read_csv('GPT_1900_EMBEDDINGS_OUTPUT_TEMP.csv')
df.set_index('doc', inplace=True)

df

def delete_files(folder_path, file_names):
    # Iterate through files in the folder
    for file_name in os.listdir(folder_path):
        file_path = path.join(folder_path, file_name)

        # Check if the file is in the list of file names (without extensions)
        if path.isfile(file_path) and path.splitext(path.basename(file_path))[0] in file_names:
            try:
                # Delete the file
                os.remove(file_path)
                print(f"File {file_name} deleted successfully.")
            except Exception as e:
                print(f"Error deleting file {file_name}: {e}")

delete_files(path_to_N_random_chunks, list(df.index))
