==> dev/01_get_gpt_labels.ipynb <==
import pandas as pd
import numpy as np
from openai import OpenAI
import joblib
from tqdm.notebook import tqdm
from glob import glob
import pickle
from os import path
from unicodedata import normalize

path_name = '/home/crazyjeannot/Documents/doctorat/AVENTURES/corpus/1000_chunks_txt/*.txt'
path_test = '/home/crazyjeannot/Documents/doctorat/AVENTURES/corpus/test_chunks_txt/*.txt'
path_chunk = '/home/crazyjeannot/Documents/doctorat/AVENTURES/corpus/all_chunks_txt/1908_Rolland-Romain_Jean-Christophe_Tome-VII-Dans-la-maison.txt_chunk_43.txt'

def clean_text(txt):
    txt_res = normalize("NFKD", txt).replace('\xa0', ' ')
    txt_res = txt_res.replace("\\", "").replace('\\xa0', '')
    return txt_res

def gpt_stuff(chunk):

    client = OpenAI(
        api_key="[OPEN-AI-KEY]",
    )

    instruction = "Give me as an output only one word, if this text is typical of the adventure novels genre; write ADVENTURE; elif unsure; write NON_ADVENTURE. Prefer NON_ADVENTURE when unsure. HELP : adventure novels are characterized by the importance of change of scenery (historical/geographical/fantastical or social) and of violent action putting the hero in mortal danger or physical peril. Typical adventure scene : someone (described as brave/heroic) doing something dangerous in a heroic manner in a wild setting."

    response = client.chat.completions.create(
        messages=[
            {"role": "system", "content": instruction},
            {"role": "user", "content": chunk}
        ],
        model="gpt-3.5-turbo-1106",
        temperature=0.2,
    )
    
    if response is not None:
        completion = response.choices[0].message.content
        return completion
    else:
        print(f'Error: Failed to get response')
        return None

def get_labels(path_name):
    df = pd.DataFrame(columns=['doc', 'label'])

    for doc in tqdm(glob(path_name)):
        doc_name = path.splitext(path.basename(doc))[0]
        with open(doc, encoding="utf8") as file:
            text = file.readlines()
            text_cleaned = clean_text(str(text).lower())

            label = gpt_stuff(text_cleaned)

            df = pd.concat([df, pd.DataFrame({'doc': [doc_name], 'label': [label]})])
            df_temp = df.set_index('doc')
            df_temp.to_csv('GPT_ANNOT_OUTPUT_TEMP.csv')
    df.set_index('doc', inplace=True)
    df.to_csv('GPT_ANNOT_OUTPUT_MAIN.csv')

    return df


df_labels = get_labels(path_name)

df_labels

df_labels.label.value_counts()
