==> dev/04_classification_distillation.ipynb <==
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm 

from sklearn.ensemble import RandomForestClassifier, HistGradientBoostingClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, accuracy_score
from sklearn.preprocessing import Normalizer, StandardScaler
from sklearn.pipeline import make_pipeline
from sklearn.svm import SVC

df = pd.read_csv('/home/crazyjeannot/Documents/doctorat/AVENTURES/data/metadata/GPT_ANNOT_OUTPUT_MAIN.csv')
df.set_index('doc', inplace=True)

def classification(df, N_sample=10):
    all_accuracy = []
    for i in tqdm(range(N_sample)):
        X_train, X_test, y_train, y_test = train_test_split(
            df.drop(['label'], axis=1), df['label'], test_size=0.2#, random_state=42
        )

        pipe = make_pipeline(StandardScaler(), SVC(probability=True))

        pipe.fit(X_train, y_train) # clf
        preds = pipe.predict(X_test)
        probas = pipe.predict_proba(X_test)

        accuracy = accuracy_score(y_test, preds)
        report = classification_report(y_test, preds)
        #print(report)
        print(accuracy)
        all_accuracy.append(accuracy)
    print(sum(all_accuracy)/len(all_accuracy))
    return pipe, accuracy

# EMBEDDINGS ALL

df_all_embeddings = pd.read_csv('/home/crazyjeannot/Documents/doctorat/AVENTURES/data/GPT_EMBEDDINGS_OUTPUT_MAIN.csv')
df_all_embeddings.set_index('doc', inplace=True)

df_merged = pd.merge(df_all_embeddings, df, left_index=True, right_index=True, how='left')

pipe, acc = classification(df_merged)

### DOC2VEC EMBEDDINGS

df_all_embeddings = pd.read_csv('/home/crazyjeannot/Documents/doctorat/AVENTURES/data/DOC2VEC_EMBEDDINGS_OUTPUT_MAIN.csv')
df_all_embeddings.set_index('doc', inplace=True)

df_merged = pd.merge(df_all_embeddings, df, left_index=True, right_index=True, how='left')

pipe, acc = classification(df_merged)

## Get Vingt Mille Lieux sous la mer - inferences

df_main = pd.read_csv('GPT_verne_EMBEDDINGS_OUTPUT_MAIN.csv')
df_main.set_index(['doc'], inplace=True)

preds = pipe.predict(df_main)
probas = pipe.predict_proba(df_main)

proba_A, proba_nonA = [proba[0] for proba in probas], [proba[1] for proba in probas]

zipped = list(zip(list(df_main.index), preds, proba_A, proba_nonA))

df_verne = pd.DataFrame(zipped, columns=['Index', 'Preds', 'Proba_A', 'Proba_nonA'])

df_verne.to_csv('preds_Vingt_Milles_Lieux.csv', index=False, header=True)

## Get l'education sentimentale - inferences

df_main = pd.read_csv('GPT_flaubert_EMBEDDINGS_OUTPUT_MAIN.csv')
df_main.set_index(['doc'], inplace=True)

preds = pipe.predict(df_main)
probas = pipe.predict_proba(df_main)

proba_A, proba_nonA = [proba[0] for proba in probas], [proba[1] for proba in probas]

zipped = list(zip(list(df_main.index), preds, proba_A, proba_nonA))

df_flaubert = pd.DataFrame(zipped, columns=['Index', 'Preds', 'Proba_A', 'Proba_nonA'])

df_flaubert.to_csv('preds_Education.csv', index=False, header=True)

### Get 10000 random chunks score

df_main = pd.read_csv('../data/EMBEDDINGS_CHAPITRES_RANDOM.csv')
df_main.set_index(['doc'], inplace=True)

preds = pipe2.predict(df_main)
probas = pipe2.predict_proba(df_main)

proba_A, proba_nonA = [proba[0] for proba in probas], [proba[1] for proba in probas]

zipped = list(zip(list(df_main.index), preds, proba_A, proba_nonA))

df_10000 = pd.DataFrame(zipped, columns=['Index', 'Preds', 'Proba_A', 'Proba_nonA'])

df_10000.to_csv('preds_10000.csv', index=False, header=True)

df_10000.Preds.value_counts()

# EMBEDDINGS BOOKNLP

df_booknlp = pd.read_csv('../data/GPT_EMBEDDINGS_1000_BOOKNLP_OUTPUT_MAIN.csv')
df_booknlp.set_index('doc', inplace=True)

df_merged_booknlp = pd.merge(df_booknlp, df, left_index=True, right_index=True, how='left')

pipe_booknlp, acc_booknlp = classification(df_merged_booknlp)

# EMBEDDINGS WITHOUT BOOKNLP

df_non_booknlp = pd.read_csv('../data/GPT_EMBEDDINGS_1000_NONBOOKNLP_OUTPUT_MAIN.csv')
df_non_booknlp.set_index('doc', inplace=True)

df_merged_non_booknlp = pd.merge(df_non_booknlp, df, left_index=True, right_index=True, how='left')

pipe_non_booknlp, acc_non_booknlp = classification(df_merged_non_booknlp)

# EMBEDDINGS 150 tokens WITHOUT BOOKNLP

df_random_non_booknlp = pd.read_csv('GPT_EMBEDDINGS_1000_RANDOM_NONBOOKNLP_OUTPUT_MAIN.csv')
df_random_non_booknlp.set_index('doc', inplace=True)

df_merged_random_non_booknlp = pd.merge(df_random_non_booknlp, df, left_index=True, right_index=True, how='left')

pipe_random_non_booknlp, acc_random_non_booknlp = classification(df_merged_non_booknlp)

GPT_EMBEDDINGS_1000_RANDOM_NONBOOKNLP_OUTPUT_MAIN
